requirements.txt
django-heroku==0.3.1
django-widget-tweaks==1.4.8
prettyconf==2.1.0
